+++ packages/example/lib/home_screen.dart
@@ -24,7 +24,12 @@ class HomeScreenState extends State<HomeScreen> {
     RudderTraits traits = RudderTraits()
         .putName("Sai Venkat")
         .putAge("22")
-        .putEmail("saivenkatdesu@gmail.com");
+        .put("city", "Hyderabad")
+        .put("state", "Telangana")
+        .putValue({"key1": "value1", "key2": "value2"}).put("details", {
+      "hobby": "football",
+      "fav_color": "red"
+    }).putEmail("saivenkatdesu@gmail.com");
     rudderClient.identify("161FA04009", traits: traits);
     setOutput(
         "identify : \nname:Sai Venkat\nage: 22\nemail:saivenkatdesu@gmail.com"
@@ -42,6 +47,7 @@ class HomeScreenState extends State<HomeScreen> {
         dotenv.env['DATA_PLANE_URL'] ?? "https://hosted.rudderlabs.com");
     builder.withMobileConfig(mc);
     builder.withLogLevel(RudderLogger.VERBOSE);
+    builder.withDataResidencyServer(DataResidencyServer.US);
     String writeKey = dotenv.env['WRITE_KEY'] ?? "INVALID_WRITE_KEY";
     rudderClient.initialize(writeKey, config: builder.build(), options: null);
     isInitialized = true;
@@ -67,6 +73,14 @@ class HomeScreenState extends State<HomeScreen> {
     RudderOption options = RudderOption();
     options.putIntegration("All", true);
     options.putIntegration("Mixpanel", false);
+    options.putCustomContext("address", {
+      "city": "kolkata",
+      "pin": "700091",
+      "state": {"name": "West Bengal", "code": "WB"},
+      "country": {"name": "India", "code": "IN"},
+      "zone": 12,
+      "lat": 22.5726,
+    });
     rudderClient.track("Went on a drive web",
         properties: property, options: options);
 
@@ -90,6 +104,8 @@ class HomeScreenState extends State<HomeScreen> {
     RudderTraits groupTraits = RudderTraits();
     groupTraits.put("place", "kolkata");
     groupTraits.put("size", "fifteen");
+    groupTraits.put("details", {"domain": "SDK", "type": "flutter"});
+    groupTraits.putValue({"key1": "value1", "key2": "value2"});
     rudderClient.group("Integrations-Rudder", groupTraits: groupTraits);
     setOutput(
         "group\n\ttraits:\n\t\tplace:kolkata\n\t\tsize:fifteen\n\tid: Integrations-Rudder");