+++ packages/plugins/rudder_plugin_android/android/src/main/java/com/rudderstack/sdk/flutter/RudderSdkFlutterPlugin.java
@@ -18,6 +18,7 @@ import io.flutter.plugin.common.MethodChannel.MethodCallHandler;
 import io.flutter.plugin.common.MethodChannel.Result;
 import java.lang.*;
 import java.util.*;
+import com.rudderstack.android.sdk.core.RudderDataResidencyServer;
 
 /** RudderSdkFlutterPlugin */
 public class RudderSdkFlutterPlugin implements FlutterPlugin, MethodCallHandler {
@@ -229,6 +230,10 @@ public class RudderSdkFlutterPlugin implements FlutterPlugin, MethodCallHandler 
         .withTrackLifecycleEvents((Boolean) configMap.get("trackLifecycleEvents"))
         .withRecordScreenViews((Boolean) configMap.get("recordScreenViews"))
         .withControlPlaneUrl((String) configMap.get("controlPlaneUrl"));
+    String dataResidencyServer = (String) configMap.get("dataResidencyServer");
+    if (dataResidencyServer.equals("EU")) {
+      builder.withDataResidencyServer(RudderDataResidencyServer.EU);
+    }
     if (integrationList != null) {
       builder.withFactories(integrationList);
     }
@@ -337,6 +342,13 @@ public class RudderSdkFlutterPlugin implements FlutterPlugin, MethodCallHandler 
       Map<String, Object> integrationsMap = (Map<String, Object>) optionsMap.get("integrations");
       for (Map.Entry<String, Object> entry : integrationsMap.entrySet()) {
         option.putIntegration(entry.getKey(), (boolean) entry.getValue());
+      }
+    }
+
+    if (optionsMap.containsKey("customContexts")) {
+      Map<String, Object> customContextsMap = (Map<String, Object>) optionsMap.get("customContexts");
+      for (Map.Entry<String, Object> customContext : customContextsMap.entrySet()) {
+        option.putCustomContext(customContext.getKey(), (Map<String, Object>) customContext.getValue());
       }
     }
     return option;