+++ packages/plugins/rudder_plugin_interface/test/src/models/rudder_config_test.dart
@@ -1,6 +1,6 @@
-import 'package:rudder_sdk_flutter_platform_interface/src/models/rudder_integration.dart';
+import 'test_integration.dart';
 import 'package:test/test.dart';
-import 'package:rudder_sdk_flutter_platform_interface/src/models/rudder_config.dart';
+import 'package:rudder_sdk_flutter_platform_interface/platform.dart';
 
 ///use flutter test
 void main() {
@@ -17,12 +17,12 @@ void main() {
         cookieConsentManagers: {"oneTrust": false});
 
     final List<RudderIntegration> factories = [
-      _create(() {}, "1"),
-      _create(() {}, "2"),
-      _create(() {}, "3"),
-      _create(() {}, "4"),
+      create(() {}, "1"),
+      create(() {}, "2"),
+      create(() {}, "3"),
+      create(() {}, "4"),
     ];
-    final RudderIntegration factory = _create(() => () {}, "5");
+    final RudderIntegration factory = create(() => () {}, "5");
 
     RudderConfig config = RudderConfigBuilder()
         .withControlPlaneUrl(configUrl)
@@ -53,21 +53,26 @@ void main() {
     expect(queueOpts["maxRetryDelay"], webConfig.maxRetryDelay.toString());
     expect(queueOpts["maxAttempts"], webConfig.maxAttempts.toString());
   });
-}
 
-RudderIntegration _create(Function() addFactory, String key) {
-  return TestIntegration(addFactory, key);
-}
+  test('data residency value is properly set', () {
+    RudderConfig config = RudderConfigBuilder()
+        .withFlushQueueSize(12)
+        .withControlPlaneUrl("https://api.dev.rudderlabs.com")
+        .withDataResidencyServer(DataResidencyServer.EU)
+        .build();
+    Map<String, dynamic> webMap = config.toMapWeb();
+    Map<String, dynamic> mobileMap = config.toMapMobile();
+    expect(webMap["residencyServer"], equals("EU"));
+    expect(mobileMap["dataResidencyServer"], equals("EU"));
 
-class TestIntegration extends RudderIntegration {
-  final Function() _addFactory;
-  final String _key;
-
-  TestIntegration(this._addFactory, this._key);
-  @override
-  String getKey() => _key;
-  @override
-  void addFactory() {
-    _addFactory.call();
-  }
+    config = RudderConfigBuilder()
+        .withFlushQueueSize(12)
+        .withControlPlaneUrl("https://api.dev.rudderlabs.com")
+        .withDataResidencyServer(DataResidencyServer.US)
+        .build();
+    webMap = config.toMapWeb();
+    mobileMap = config.toMapMobile();
+    expect(webMap["residencyServer"], equals("US"));
+    expect(mobileMap["dataResidencyServer"], equals("US"));
+  });
 }